examples/slack-bot/requirements.txt
# /malloy-slack-bot/requirements.txt

# Slack Integration
slack_sdk
python-dotenv

# LLM Integration
openai
anthropic

# HTTP Clients & Web Framework
requests  # Legacy - will migrate to aiohttp
aiohttp   # New async HTTP client
flask     # Health check endpoints for cloud deployment

# Data Processing & Visualization
pandas
matplotlib

# Async Support
asyncio  # Built-in, but listed for clarity

# LangChain Integration - Phase 2
langchain
langchain-openai
langchain-community
langchain-google-vertexai
pydantic

# Testing & Development
pytest
pytest-asyncio
pytest-mock
pytest-cov
black
flake8
mypy